serialNum

+++ shared/src/main/scala/org/seekloud/thor/shared/ptcl/thor/EsRecover.scala
@@ -86,7 +86,7 @@ trait EsRecover {
     if (rollBackFrame.contains(frame)) rollBackFrame = rollBackFrame.filterNot(_ == frame)
   }
 
-  def removePreEventHistory(frame: Int, playerId: Short, serialNum: Int): Unit = {
+  def removePreEventHistory(frame: Int, playerId: Short, serialNum: Byte): Unit = {
     actionEventHistoryMap.get(frame).foreach { actions =>
       actionEventHistoryMap.put(frame, actions.filterNot(t => t.playerId == playerId && t.serialNum == serialNum))
     }
